+++ kurtosis-module/impl/participant_network/cl/cl_client_launcher.go
@@ -3,7 +3,7 @@ package cl
 import (
 	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/el"
 	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/log_levels"
-	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/prelaunch_data_generator"
+	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/prelaunch_data_generator/cl"
 	"github.com/kurtosis-tech/kurtosis-core-api-lib/api/golang/lib/enclaves"
 	"github.com/kurtosis-tech/kurtosis-core-api-lib/api/golang/lib/services"
 )
@@ -17,7 +17,7 @@ type CLClientLauncher interface {
 		// If nil, the node will be launched as a bootnode
 		bootnodeContext *CLClientContext,
 		elClientContext *el.ELClientContext,
-		nodeKeystoreDirpaths *prelaunch_data_generator.NodeTypeKeystoreDirpaths,
+		nodeKeystoreDirpaths *cl.NodeTypeKeystoreDirpaths,
 	) (
 		resultClientCtx *CLClientContext,
 		resultErr error,