+++ kurtosis-module/impl/prelaunch_data_generator/el_prelaunch_data_generator.go
@@ -1,0 +1,4 @@
+package prelaunch_data_generator
+
+func generateElPrelaunchData() {
+}